refactor to use latest twitter typeahead
