+++ README.md
@@ -47,6 +47,13 @@ You may be on a page that's focusing on a text area or input box. If this is the
   * Use search icon instead of J as a more intuitive icon
 4. 0.31
   * Quick bug fix for icon issue that lead to database results not being shown
+5. 0.40
+  * Code maintenance
+  * Ability to search by post_id
+  * add user setting for invoking jarvis
+6. 0.50.0
+  * Refactored to use the latest Twitter Typeahead
+
 
 ## Authors
