+++ css/jarvis.css
@@ -6,13 +6,12 @@
 #jarvis-modal .tt-hint {color:rgb(150,150,150);}
 #jarvis-modal a {display:block;text-decoration: none;padding:.5em;border:1px solid rgb(255,255,255);}
 .twitter-typeahead {width:100%;}
-.ie8 .twitter-typeahead {background:rgb(235,235,235);}
-.tt-dropdown-menu {background:rgb(255,255,255);border-radius:.5em;padding:1em;margin:0 -1em 0 -1em;width:100%;}
+.tt-menu {background:rgb(255,255,255);border-radius:.5em;padding:1em;margin:0 -1em 0 -1em;width:100%;}
 .tt-suggestion {overflow:hidden;}
 .tt-suggestion p {margin:0;}
 .tt-suggestion a:hover,
-.tt-is-under-cursor a {border-radius:.5em;background:rgb(239, 248, 255);color:rgb(213,78,33);}
-.tt-is-under-cursor a:hover {border-color:rgb(209,218,225);}
+.tt-cursor a {border-radius:.5em;background:rgb(239, 248, 255);color:rgb(213,78,33);}
+.tt-cursor a:hover {border-color:rgb(209,218,225);}
 .jarvis-icon {background:url(../../../../wp-includes/images/admin-bar-sprite.png) 0 -100px no-repeat;line-height:28px;width:28px;height:28px;float:left;}
 .jarvis-icon.dashicons-before {background:none;}
 .jarvis-icon.dashicons-before:before {vertical-align:middle;}
