+++ package.json
@@ -1,0 +1,30 @@
+{
+  "name": "jarvis",
+  "version": "0.50.0",
+  "description": " Jarvis is your admin assistant, putting WordPress at your fingertips via a quicksearch interface.",
+  "main": "index.js",
+  "scripts": {
+    "gulp": "gulp"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/WDGDC/Jarvis.git"
+  },
+  "author": "WDGDC",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/WDGDC/Jarvis/issues"
+  },
+  "homepage": "https://github.com/WDGDC/Jarvis#readme",
+  "devDependencies": {
+    "gulp": "^3.9.1",
+    "gulp-concat": "^2.6.0",
+    "gulp-header": "^1.7.1",
+    "gulp-insert": "^0.5.0",
+    "gulp-uglify": "^1.5.3"
+  },
+  "dependencies": {
+    "hogan.js": "^3.0.2",
+    "typeahead.js": "^0.11.1"
+  }
+}
