+++ readme.txt
@@ -73,7 +73,10 @@ You may be on a page that's focusing on a text area or input box. If this is the
 = 0.31 =
 * Quick bug fix for icon issue that lead to database results not being shown
 
-= 0.4 =
+= 0.40 =
 * Code maintenance
 * Ability to search by post_id
 * add user setting for invoking jarvis
+
+= 0.50.0 =
+* Refactored to use the latest Twitter Typeahead
